Improve the robustness of the server side event processing
We are now processing the events in the order of their kind:
- First we process all the joins, sending out all the necessary
greetings and notifying other players about the joiners,
- Then we process all the leaves,
- And only then notifying about movements.

This ensures correct order of sent out events, and filters out noise
like for instance somebody joining, moving a bit and leaving very
quickly within a single tick.

We also added the entirety of the information about the player in the
'Hello' message, so the clients know about themselves as soon as they
are greeted. Because of that we are trying not to send out
'PlayerJoined' event to the players about themselves, because that's
redundant now.